作成中

+++ include/Tester/mpc/ltv_mpc/servo_motor_ltv_constraints_Y_min.hpp
@@ -7,18 +7,12 @@ namespace servo_motor_ltv_constraints_Y_min {
 
 using namespace PythonNumpy;
 
-using SparseAvailable_Y_min = SparseAvailable<
-    ColumnAvailable<false>,
-    ColumnAvailable<false>
->;
+using SparseAvailable_Y_min =
+    SparseAvailable<ColumnAvailable<false>, ColumnAvailable<false>>;
 
-using type = SparseMatrix_Type<double, SparseAvailable_Y_min>;
+using type = SparseMatrix_Type<float, SparseAvailable_Y_min>;
 
-inline auto make(void) -> type {
-
-  return make_SparseMatrixEmpty<double, 2, 1>();
-
-}
+inline auto make(void) -> type { return make_SparseMatrixEmpty<float, 2, 1>(); }
 
 } // namespace servo_motor_ltv_constraints_Y_min
