+++ include/Tester/mpc/ltv_mpc/servo_motor_ltv_constraints_ltv_mpc_Weight_U_Nc.hpp
@@ -7,15 +7,12 @@ namespace servo_motor_ltv_constraints_ltv_mpc_Weight_U_Nc {
 
 using namespace PythonNumpy;
 
-using type = DiagMatrix_Type<double, 2>;
+using type = DiagMatrix_Type<float, 2>;
 
 inline auto make(void) -> type {
 
-  return make_DiagMatrix<2>(
-    static_cast<double>(0.001),
-    static_cast<double>(0.001)
-  );
-
+  return make_DiagMatrix<2>(static_cast<float>(0.001),
+                            static_cast<float>(0.001));
 }
 
 } // namespace servo_motor_ltv_constraints_ltv_mpc_Weight_U_Nc
