+++ include/Tester/mpc/ltv_mpc/servo_motor_ltv_constraints_ltv_mpc_lkf_ss_A.hpp
@@ -7,30 +7,24 @@ namespace servo_motor_ltv_constraints_ltv_mpc_lkf_ss_A {
 
 using namespace PythonNumpy;
 
-using SparseAvailable_ltv_mpc_lkf_ss_A = SparseAvailable<
-    ColumnAvailable<true, true, false, false>,
-    ColumnAvailable<true, true, true, false>,
-    ColumnAvailable<false, false, true, true>,
-    ColumnAvailable<true, false, true, true>
->;
+using SparseAvailable_ltv_mpc_lkf_ss_A =
+    SparseAvailable<ColumnAvailable<true, true, false, false>,
+                    ColumnAvailable<true, true, true, false>,
+                    ColumnAvailable<false, false, true, true>,
+                    ColumnAvailable<true, false, true, true>>;
 
-using type = SparseMatrix_Type<double, SparseAvailable_ltv_mpc_lkf_ss_A>;
+using type = SparseMatrix_Type<float, SparseAvailable_ltv_mpc_lkf_ss_A>;
 
 inline auto make(void) -> type {
 
   return make_SparseMatrix<SparseAvailable_ltv_mpc_lkf_ss_A>(
-    static_cast<double>(1.0),
-    static_cast<double>(0.05),
-    static_cast<double>(-2.5603853840856567),
-    static_cast<double>(0.9500002466138069),
-    static_cast<double>(0.12801926920428283),
-    static_cast<double>(1.0),
-    static_cast<double>(0.05),
-    static_cast<double>(6.4009950316892),
-    static_cast<double>(-0.32004975158446064),
-    static_cast<double>(0.4900000000000001)
-  );
-
+      static_cast<float>(1.0), static_cast<float>(0.05),
+      static_cast<float>(-2.5603853840856567),
+      static_cast<float>(0.9500002466138069),
+      static_cast<float>(0.12801926920428283), static_cast<float>(1.0),
+      static_cast<float>(0.05), static_cast<float>(6.4009950316892),
+      static_cast<float>(-0.32004975158446064),
+      static_cast<float>(0.4900000000000001));
 }
 
 } // namespace servo_motor_ltv_constraints_ltv_mpc_lkf_ss_A
